fix databse typo in test imports to match renamed database folder
