+++ test/data/database/SmallItemSqlTableDao.test.ts
@@ -1,5 +1,5 @@
 import { expect } from 'vitest';
-import { SmallItemSqlTableDao } from '../../../src/data/databse/SmallItemSqlTableDao';
+import { SmallItemSqlTableDao } from '../../../src/data/database/SmallItemSqlTableDao';
 import type { SmallItem } from '../../../src/domain/models/items/SmallItem';
 import type { EquipmentFilters } from '../../../src/domain/models/items/EquipmentFilters';
 import { runSqlDaoBaseTests } from './Dao';
